Refactor venues to templates
