+++ src/components/organisms/TemplateRouter/TemplateRouter.tsx
@@ -3,8 +3,8 @@ import { useParams } from "react-router-dom";
 import useConnectCurrentVenue from "hooks/useConnectCurrentVenue";
 import { Venue } from "types/Venue";
 import { VenueTemplate } from "types/VenueTemplate";
-import EntranceExperience from "components/venues/Jazzbar/EntranceExperience";
-import PartyMapRouter from "components/venues/PartyMap/router";
+import EntranceExperience from "components/templates/Jazzbar/EntranceExperience";
+import PartyMapRouter from "components/templates/PartyMap";
 import { ChatContextWrapper } from "components/context/ChatContext";
 import { useUser } from "hooks/useUser";
 import { useSelector } from "hooks/useSelector";
