+++ src/components/templates/Jazzbar/JazzBarSkeletonPage/JazzBarSkeletonPage.scss
