+++ src/components/templates/Jazzbar/JazzBarSkeletonPage/JazzBarSkeletonPage.tsx
