+++ src/components/templates/PartyMap/RoomPage/RoomPage.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import { getCurrentEvent } from "utils/time";
 import Chatbox from "components/organisms/Chatbox";
-import RoomModalOngoingEvent from "components/venues/PartyMap/components/RoomModalOngoingEvent";
+import RoomModalOngoingEvent from "components/templates/PartyMap/components/RoomModalOngoingEvent";
 import UserList from "components/molecules/UserList";
-import ScheduleItem from "components/venues/PartyMap/components/ScheduleItem";
+import ScheduleItem from "components/templates/PartyMap/components/ScheduleItem";
 import WithNavigationBar from "components/organisms/WithNavigationBar";
 import { enterRoom } from "utils/useLocationUpdateEffect";
 import { updateTheme } from "pages/VenuePage/helpers";
