+++ src/components/templates/PartyMap/components/Map/index.js
