+++ src/components/templates/PartyMap/components/PartyTitle/PartyTitle.scss
