+++ src/components/templates/PartyMap/components/PartyTitle/index.ts
