+++ src/components/templates/PartyMap/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PartyMapRouter";
