--- src/components/venues/PartyMap/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./router";
