+++ src/pages/ReactionPage/ReactionPage.tsx
@@ -4,7 +4,7 @@ import { useFirestoreConnect } from "react-redux-firebase";
 import useConnectPartyGoers from "hooks/useConnectPartyGoers";
 import "./ReactionPage.scss";
 import UserList from "components/molecules/UserList";
-import ReactionList from "components/venues/Jazzbar/components/ReactionList";
+import ReactionList from "components/templates/Jazzbar/components/ReactionList";
 import { useSelector } from "hooks/useSelector";
 import { MessageToTheBandReaction } from "components/context/ExperienceContext";
 import { WithId } from "utils/id";
